Removing logging calls from signal handlers
Seeing an issue in production where this signal handler is being handled
in a way that causes the eventlet hub to *not* be able to properly switch back
to its runloop properly.

My hunch is that this is happening due to the use of the threadpool and signal
handlers being forced to execute on the primary thread. That said, no repro
case whatsoever. HEISENBUG!

Change-Id: Id303e3755d89d7e2aab06dde17719fef843b01a2